Get basic routing with react-router-dom working
This requires us to serve the same index.html page, regardless of the page
the user actually navigates to. At first, this seemed really difficult, but
a feature of WebPack (setting `publicPath` to `/static/`) allows us to
easily detect which assets should actually be served (e.g. for JS scripts
etc.), and which should be rerouted to the index page.