Don't use Yarn zero-installs
Pushing every single dependency to the repo just doesn't sit right with me:
the repo is enormous for how much code we actually have, the history is
messy and filled with files you should just ignore, and reading the benefits
Yarn says it has is... underwhelming. This hopefully rewrites the history
to not do that anymore.