Write a format to represent the content of Zettels
This is a really basic representation of some of the things our Zettels
will be able to contain - hopefully enough to play around with our editor
with. We send it between the frontend and backend as JSON - if this becomes
problematic in the future we can explore a more elaborate update-tracking
system, or a more concise format / wire format.