Improve paragraph spacing a little
It's proving a little tricky to get paragraphs spaced nicely with my very
limited knowledge of CSS. We basically want paragraphs to not have huge
amounts of spacing around them like they did before, but also look slightly
separated when paragraphs are situated next to each other with nothing in
between. This probably needs a bunch more tweaking but looks okay for now.