Add component to report changes to the Prosemirror document
This allows `ZettelEditor` to provide a function that is called whenever
the contents of the Prosemirror editor is changed, debounced to avoid
loads of calls when the user is actively typing. We can then use this to
update the state of the Zettel.

This also required us to implement a basic serializer, which walks the
Prosemirror contents and constructs a new version of our representation.