Cache Zettels on the frontend
This uses React context to cache the Zettels we've fetched from the server,
so we don't have to fetch them over and over again, especially for things
like just creating links to them in the sidebar. I've used this newfangled
thing called a reducer for this, as I think the cache will increase a fair
bit in complexity - this might be overkill / not the right solution though,
in which case we can rip it out.