Handle both internal and external links being clicked
Because we're using `react-router`, we need to detect internal URLs and
use the History API to open them. At the moment, we just do this if the URL
begins with an '/'. External URLs are opened in a new tab to avoid losing
changes to the Zettel atm (and because I think it's a better behaviour
overall).