Don't bother fetching a Zettel if it's already in the cache
This does present some problems with letting the cache get stale: what if
we change a Zettel from one device and update it, but this change isn't
reflected on another device because it still reckons it has the Zettel
cached. I guess in the future the server will need some way of invalidating
cached Zettels (maybe tracking the Zettels a device has cached? idk).