Add a functionality to open a menu when '/' is typed
I thought this was going to be really tricky to do, but I found a nice
technique in the Milkdown editor (https://github.com/Saul-Mirone/milkdown/tree/main/packages/plugin-slash)
that works well.

Basically, when the user types '/', we call a function which will open a menu.
Placeholder text is placed using Prosemirror node decorations that tell the
user they can do this (and also nicely mark where the editor actually is
when it's empty).