Add some buttons to the insert menu that insert headers
This is s proof of concept of the insert menu - it adds four buttons that
each insert a header of a different level. This tests out that we're able
to issue Prosemirror commands, and that the command constructor idea works
(it sort of does but is fairly verbose with the casting). There are still
some rough edges (we need to get rid of the '/' when a command actually gets
done, and we need some actual styling).