Split Zettel format into two: record and endpoint formats
We're splitting the format that Zettels are stored in from the one that's
served on JSON endpoints, for slightly annoying but practical reasons. For
a start, this is required for us to move to Bincode, as the internal tagging
that we use in the JSON format is incompatible with non-self-describing
formats.