Store Zettels in bincode and add version tag
We prepend each Zettel with a version tag to make migration between formats
possible in the future. Our first format will be encoded with bincode, which
does seem to make a measurable difference in how much space Zettels use.

To make this possible, we separate out the `ZettelStore` structure from the
endpoints, as we now need to convert between the endpoint and record formats
at the barrier.