Rewrite `insertHeader` command to avoid bug with insert menu
I'm not sure if this is the right way to fix this (as in whether the original
command should work, or WAS the bug), but before, this command would
convert the preceding block to a header, rather than the current one, because
we remove the `/` used to trigger the menu first. This was modelled on how
we insert dividers, and seems to fix that.