Only update Zettel title and content if dirty
This tracks whether the title and content of a Zettel are actually dirty and
need to be sent with an update. This will reduce the number of updates we
send, and avoids sending the entire content when the title changes (which
isn't actually a biggy but made me feel nasty).

This isn't perfect, for reasons(TM) `onChange` and `onChangeTitle` get called
at the beginning when we set them from the fetched Zettel, which marks them
as dirty. I've decided this doesn't matter for now, as this is still an
improvement, and also allows us to implement autosave without feeling super
nasty.