Simplify Zettel saving by just offloading it to the backend
This replaces all of the special-casing we've done of Zettel saving we've
done so far with a very simple solution - just send it to the server all
the time and let it deal with it. This actually works better than expected,
and also effectively implements autosaving.

The only remaining problem is effectively I think one with the server - it
takes a fairly long time to actually respond to the update (e.g. a second,
whereas something like Notion takes a few hundred ms even over the Internet).
I think this is just because we block on saving and re-indexing the Zettel
on every single update - we should probably schedule this onto some sort of
worker thread pool and respond immediately.