Commit the Tantivy index in a background task
Tantivy has a significant problem: committing the index is reallllly slow.
It pulls a Zettel update from 2ms to around 1300ms, which is fairly ludicrous.
I'm assuming this is because this is not really the usecase of Tantivy: I
think it expects you to index a whole heap of stuff and then commit once, and
then basically not want to change that indexed stuff again. This is kind of
the polar opposite to what we need :P

This "solves" this problem by offloading the actual commit to a background
task, which runs periodically, commiting the index if changes have been made
to it. This can still slow requests down: if you try and index something just
as the commit starts, you'll still have to wait for it to end and release
the index writer to you. But it's still way better than before, and also
hopefully much less computationally expensive (although it's still not great
to need to be re-indexing the Zettel on every update).

However, I think in the future we'll need to consider whether Tantivy provides
the right tradeoffs for our use case (where tbh, searching is going to be
pretty rare, and we're not even really indexing a huge amount of data, but
updating the index needs to be fairly fast and inexpensive.