Try to create stable Zettel format for now
We currently don't have tools for migrating Zettels between versions, so
atm we just delete the entire corpus when the format changes. This is fine
for testing, but we actually want to start dogfooding this at some point
so it's ready for us to actually use at some point. This format will hopefully
tide us over for now in a way that we just need to add new stuff, rather
than change it in a way that will break backwards compatability.