Fix yet another of Docker's weird intricacies
For some reason Docker's not copying across the hidden `.cargo/config` (we're
not ignoring it in .dockerignore), and so the ARM build fails because we're
trying to link with `cc`. We also can't manually copy it over, but we can
manually copy the entire directory over. We need to do this before `cargo cook`ing
btw as we, I'm guessing, are actually doing a build in there somewhere.