Temporary hack for scrollable sidebar
This is not the final version of this, but I need to be able to scroll the
sidebar so let's get this deployed lmao. This allows you to scroll the whole
bar, whereas we really want to apply it to specific sections, but this is
harder because of course it is because CSS - we'll come back to it.