Use default indent on `ul` elements
Not sure what our original intention with this was, but unordered lists
were way too subtle. It's nicer having a little bit of identation. I've
kept the `margin: 0` because otherwise the top of the list is spaced out
like a paragraph of it's own, which doesn't really fit our writing style.